`Attribute` is mostly immutable, but the deserialized value is memoized, and that value can potentially be mutable, and mutated.

When the value is immutable however, we can share instances.

Benchmark: https://gist.github.com/casperisfine/ae56bec1e7eecbff3a696b367e2bafa2

Before:

```
ruby 3.4.0preview2 (2024-10-07 master 32c733f57b) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
   assign_attributes     6.140k i/100ms
Calculating -------------------------------------
   assign_attributes     60.732k (± 1.9%) i/s   (16.47 μs/i) -    307.000k in   5.056757s
```

After:

```
ruby 3.4.0preview2 (2024-10-07 master 32c733f57b) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
   assign_attributes     8.245k i/100ms
Calculating -------------------------------------
   assign_attributes     82.410k (± 1.3%) i/s   (12.13 μs/i) -    412.250k in   5.003220s
```

The profile now look like:

<img width="2558" alt="Capture d’écran 2024-10-16 à 14 45 10" src="https://github.com/user-attachments/assets/29c096c9-efdd-4bf3-b99b-055695678a33">


FYI: @tenderlove @Stivaros